refactor: extract ConflictFrameLoader from ConflictAnalyzerService
Move frame-loading pipeline into ConflictFrameLoader:
- resolveAnalysisContext, resolveStrandContext, resolveFrontierContext
- loadFrontierPatchFrames, buildPatchFrames, buildPatchFrame
- attachReceipts, emptyReceipt
- buildResolvedCoordinate, buildResolvedStrandMetadata
- All sorting/comparison helpers and context normalization

Convert ScanWindow from typedef+builder to a proper class with
constructor logic. Convert PatchFrame from typedef to a class
(not frozen — receipt is mutated by attachReceipts).

ConflictAnalyzerService.js: 2017 → 1596 lines.
ConflictFrameLoader.js: 473 lines.
6740 tests pass.

Part of cycle 0012 phase 3: frame loader extraction.
